Create README.md
Diabetes is a disease that is caused due to excessive amount of blood sugar in it. Our body needs energy, and glucose is one of the main sources of energy to build the muscles and tissues of the body. Generally, unhealthy lifestyle and lack of exercise are the main causes of type 2 diabetes in people. The presence of a large amount of sugar in the blood causes diabetes. Sometimes, the pancreas is unable to convert the food into insulin; thus, sugar remains unabsorbed, which causes diabetes. Diabetes can affect kidney, eyes, nervous sys